+++ src/maxtext/kernels/attention/splash_attention_kernel.py
@@ -1,7 +1,7 @@
 # pylint: skip-file
 from __future__ import annotations
 
-# Copyright 2023–2025 Google LLC
+# Copyright 2023–2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
