+++ tests/unit/quantizations_test.py
@@ -35,9 +35,9 @@ from aqt.jax.v2.flax import aqt_flax
 from MaxText.globals import MAXTEXT_PKG_DIR
 from MaxText import pyconfig
 from MaxText.layers import nnx_wrappers, quantizations
-from MaxText.kernels.megablox import gmm
 from MaxText.common_types import DECODING_ACTIVE_SEQUENCE_INDICATOR
 from maxtext.common.gcloud_stub import is_decoupled
+from maxtext.kernels.megablox import gmm
 from maxtext.utils import maxtext_utils
 from maxtext.utils import model_creation_utils
 from tests.utils.test_helpers import get_test_config_path
